Use intermediate rendering texture to flip flutter content
This produces much less glitchy resizing compared to flipping via transformation on the directcomposition visual. Unfortunately, there's an extra texture + copy :(

Might be possible to switch between these methods on demand when a resize is in progress. By default we'd use the fast path, and when a reesize is initiated we switch to using an intermediate texture, reverting back after a short delay with no resizing.